requirements.txt
farm_haystack
more_itertools
pdfkit
pdfrw
PyVirtualDisplay
streamlit
txtai
txtmarker
faiss-cpu
pydantic
pandas
nltk
torch